fix(duvet-coverage): Make proofs compile under pinned Verus
- Cast (ann_a.end_line + 1) as u64 in lemma_stacking_transitivity:
  Verus proof-mode arithmetic yields int, but all_lines_skippable
  expects u64 (E0308).
- Widen annotation_target and TargetLine to pub so the public
  is_annotation_executed postcondition can reference them.

Lets cargo verus build -p duvet-coverage reach the proof obligations.
Verification still surfaces the annotation_target mode error, which is
the next item to fix; this commit is to confirm CI reproduces the same
result as local.

+++ duvet-coverage/src/proofs.rs
@@ -339,7 +339,7 @@ proof fn lemma_stacking_transitivity(
         // A is above B: A ends before B starts
         ann_a.end_line < ann_b.start_line,
         // All lines between A's end and B's end are skippable
-        all_lines_skippable(classifications, ann_a.end_line + 1, ann_b.end_line),
+        all_lines_skippable(classifications, (ann_a.end_line + 1) as u64, ann_b.end_line),
         // B is Executed
         is_annotation_executed(ann_b, classifications, scopes, coverage, file_length)
             == ExecutionStatus::Executed,
